requirements.txt
pytorch-lightning==1.6.1
Pillow==9.1.0
numpy==1.22.3
tensorboard==2.8.0
matplotlib==3.5.1
pandas==1.4.2
seaborn==0.11.2
psutil
scikit-image==0.19.2
opencv-python==4.5.5.64